% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudopresences.R
\name{pseudo.presences.sampling}
\alias{pseudo.presences.sampling}
\title{Creates pseudo presences based on other presences to increase}
\usage{
pseudo.presences.sampling(sp, resp, env, strategy = "random",
  cell.distance = 0, np.points = NULL, quant.SRE = 0)
}
\arguments{
\item{sp}{coordiantes of reponse points (2 column matrix)}

\item{resp}{Response Variable (monospecific) as vector, 1=pres, 0=true_abs, NA=no_info}

\item{env}{Explanatory Variable as rasterStack}

\item{strategy}{Strategy to be used to generate points, one of c("random", "sre", "similar", "dissimilar")}

\item{cell.distance}{Maximum cell.distance from presence point to be used for pseudo presence generation}

\item{np.points}{Number of pseudo presence points to generate}
}
\value{
a SpatialPointsDataFrame object that will be given to BIOMOD_FormatingData function
}
\description{
Creates pseudo presences based on other presences to increase
}
